[T-186] Auto discover language specific components
