+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpDelegateMethodDeclaration.razor
@@ -10,7 +10,7 @@
 
 <span class="refdocgen-text-tertiary">@string.Join(' ', Delegate.Modifiers.CSharpData)</span>
 
-<GenericTypeLink Type="@Delegate.ReturnType" />
+<CSharpGenericTypeLink Type="@Delegate.ReturnType" />
 
 <span class="refdocgen-text-primary">
     @Delegate.Name
