+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpEventDeclaration.razor
@@ -16,7 +16,7 @@
     @string.Join(' ', Event.Modifiers.CSharpData)
 </span>
 
-<GenericTypeLink Type="@Event.Type" />
+<CSharpGenericTypeLink Type="@Event.Type" />
 
 <span class="refdocgen-text-primary">
     @eventName
