+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpFieldDeclaration.razor
@@ -12,7 +12,7 @@
     @string.Join(' ', Field.Modifiers.CSharpData)
 </span>
 
-<GenericTypeLink Type="@Field.Type" />
+<CSharpGenericTypeLink Type="@Field.Type" />
 
 <span class="refdocgen-text-primary">
     @Field.Name
