+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpGenericTypeLink.razor
@@ -6,7 +6,7 @@
     public required GenericTypeLinkTM Type { get; set; }
 }
 
-<CodeLink Type="Type.TypeLink" />
+<CodeLink Type="Type.TypeLink" LanguageId="csharp-lang" />
 
 @if (Type.TypeParameters.Any())
 {
