+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpIndexerDeclaration.razor
@@ -12,7 +12,7 @@
     @string.Join(' ', Indexer.Modifiers.CSharpData)
 </span>
 
-<GenericTypeLink Type="@Indexer.Type" />
+<CSharpGenericTypeLink Type="@Indexer.Type" />
 
     @if (Indexer.Parameters.Any())
 {
