+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpMethodDeclaration.razor
@@ -17,7 +17,7 @@
     @string.Join(' ', Method.Modifiers.CSharpData)
 </span>
 
-<GenericTypeLink Type="@Method.ReturnType" />
+<CSharpGenericTypeLink Type="@Method.ReturnType" />
 
 <span class="refdocgen-text-primary">
     @methodName
