+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpOperatorDeclaration.razor
@@ -15,7 +15,7 @@
 
 @if (!Operator.IsConversionOperator)
 {
-    <GenericTypeLink Type="@Operator.ReturnType" />
+    <CSharpGenericTypeLink Type="@Operator.ReturnType" />
 }
 
 <span></span>
