+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpParameterDeclaration.razor
@@ -19,7 +19,7 @@
     <span> </span>
 }
 
-<GenericTypeLink Type="@Parameter.Type" />
+<CSharpGenericTypeLink Type="@Parameter.Type" />
 
 <span>
     @Parameter.Name
