+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpPropertyDeclaration.razor
@@ -17,7 +17,7 @@
     @string.Join(' ', Property.Modifiers.CSharpData)
 </span>
 
-<GenericTypeLink Type="@Property.Type" />
+<CSharpGenericTypeLink Type="@Property.Type" />
 
 <span class="refdocgen-text-primary">
     @propertyName
