+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/CSharpTypeParameterConstraints.razor
@@ -33,7 +33,7 @@
 
                 foreach (var constraint in TypeParameter.TypeConstraints)
                 {
-                    <GenericTypeLink Type="@constraint" />
+                    <CSharpGenericTypeLink Type="@constraint" />
 
                     if (constraint != lastItem)
                     {
