+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/LanguageSpecific/CSharp/Utils/ParametersString.razor
@@ -19,7 +19,7 @@
 
     foreach (var p in Parameters)
     {
-        <ParameterDeclaration Parameter="@p" />
+        <CSharpParameterDeclaration Parameter="@p" />
 
         <LiteralValueSuffix Value="@p.DefaultValue" />
