+++ src/RefDocGen/TemplateProcessors/Default/Templates/Components/Menu/BottomMenu.razor
@@ -1,4 +1,5 @@
 @inherits ComponentBase
+@using System.Text.Json
 
 @code {
     [Parameter]
