+++ src/RefDocGen/TemplateProcessors/Default/Templates/SearchPage.razor
@@ -1,6 +1,7 @@
 @using RefDocGen.TemplateProcessors.Shared.TemplateModels.Language
 @using RefDocGen.TemplateProcessors.Shared.TemplateModels.Menu
 @using RefDocGen.TemplateProcessors.Shared.TemplateModels.Types
+@using System.Text.Json
 @inherits ComponentBase
 
 @code {
