+++ src/RefDocGen/TemplateProcessors/Default/Templates/_Imports.razor
@@ -11,4 +11,3 @@
 @using RefDocGen.TemplateProcessors.Default.Templates.Components.Sections.Members
 @using RefDocGen.TemplateProcessors.Default.Templates.Components.Sections.Members.Additional
 @using RefDocGen.TemplateProcessors.Default.Templates.Layout
-@using System.Text.Json
