--- src/RefDocGen/TemplateProcessors/Shared/TemplateTools/LanguageComponent.cs
@@ -1,26 +1,0 @@
-namespace RefDocGen.TemplateProcessors.Shared.TemplateTools;
-
-#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
-
-/// <summary>
-/// Enum representing the language specific components.
-/// </summary>
-public enum LanguageSpecificComponent
-{
-    AttributeDeclaration,
-    ConstructorDeclaration,
-    DelegateMethodDeclaration,
-    EnumMemberDeclaration,
-    EventDeclaration,
-    FieldDeclaration,
-    IndexerDeclaration,
-    MethodDeclaration,
-    OperatorDeclaration,
-    ParameterDeclaration,
-    PropertyDeclaration,
-    TypeDeclaration,
-    TypeParameterDeclaration,
-    TypeParameterConstraints,
-    GenericTypeLink
-}
-#pragma warning restore CS1591
